Add feature to expand/collapse long chat messages
The update involves adding a new feature to the ChatPage component that allows users to expand or collapse messages if they exceed a certain length. Also, it removes a previously used mock variable and conditionally toggles the "Show more" or "Show less" options based on whether a chat message is expanded or not. This will help in enhancing the user experience by providing readable chunks of information.