Update turbo version and modify build scripts
The update involves upgrading the version of turbo to v2.0.1 in both 'package.json' and 'pnpm-lock.yaml' files. Additionally, 'turbo.json' and the 'build' script in the 'package.json' of the 'learnbefore' app have been formatted for improved readability. The 'name' and 'dev' and 'preview' scripts in the 'package.json' have been slightly adjusted for the 'learnbefore' app.